added-simulador, refactor-comparador, added 'GO' delimiter to liquibase stored procedure.

+++ project/workspace/cadenas/cadenas_db/src/main/resources/liquibase/tables/precioProducto.sql
@@ -1,9 +1,10 @@
 CREATE TABLE precioProducto (
     idPrecio SMALLINT IDENTITY (1,1)
-   ,precio REAL NOT NULL
+   ,precio NUMERIC (10,2) NOT NULL
    ,idSucursal SMALLINT NOT NULL
    ,codigoDeBarras VARCHAR (100) NOT NULL
    ,validoDesde DATETIME DEFAULT GETDATE()
    ,PRIMARY KEY (idPrecio)
-   ,FOREIGN KEY (idSucursal,codigoDeBarras) REFERENCES producto_sucursal (idSucursal,codigoDeBarras)
+   ,FOREIGN KEY (idSucursal,codigoDeBarras)
+    REFERENCES producto_sucursal (idSucursal,codigoDeBarras)
 )
